Bug squashing & Fixing intended mechanics
Fixed an issue causing the leader party being included in influence army cost tooltips.
Intended mechanic for being able to disband an army even when you are in that army added from the kingdom army screen.
Fixed an issue where the incorrect amount of influence would be returned to a previous army leader after the player forces leaderships changed.